requirements.txt
appdirs==1.4.3
click==6.7
six==1.10.0
django-chartit==0.2.9
coreapi==2.3.0
coreapi-cli==1.0.6
coreschema==0.0.4
Django==1.8.2
dj-database-url==0.4.2
djangorestframework==3.6.2
gunicorn==19.7.1
httpie==0.9.9
itypes==1.1.0
Jinja2==2.9.6
lxml==3.7.3
MarkupSafe==1.0
numpy==1.12.1
packaging==16.8
psycopg2==2.7
Pygments==2.2.0
pyparsing==2.2.0
pyquery==1.2.10
requests==2.13.0
uritemplate==3.0.0
whitenoise==3.3.0
